Move remaining chain objects from `pkg/coordinator` to `pkg/tbtc`
For now, we agreed that all chain objects should live in the `pkg/tbtc`
which is the core package representing the tbtc domain.
